+++ src/dcos_e2e_cli/dcos_aws/commands/_options.py
@@ -6,6 +6,8 @@ from typing import Callable
 
 import click
 
+from dcos_e2e.backends import AWS
+
 from ._common import LINUX_DISTRIBUTIONS
 
 
@@ -14,10 +16,11 @@ def aws_instance_type_option(command: Callable[..., None],
     """
     An option decorator for AWS instance types.
     """
+    default_instance_type = AWS().aws_instance_type
     function = click.option(
         '--aws-instance-type',
         type=str,
-        default='m4.large',
+        default=default_instance_type,
         show_default=True,
         help='The AWS instance type to use.',
     )(command)  # type: Callable[..., None]
@@ -28,10 +31,12 @@ def aws_region_option(command: Callable[..., None]) -> Callable[..., None]:
     """
     An option decorator for AWS regions.
     """
+    default_region = AWS().aws_region
+
     function = click.option(
         '--aws-region',
         type=str,
-        default='us-west-2',
+        default=default_region,
         show_default=True,
         help='The AWS region to use.',
     )(command)  # type: Callable[..., None]