+++ tests/test_dcos_e2e/test_node_install.py
@@ -8,7 +8,6 @@ from textwrap import dedent
 from dcos_e2e.backends import Docker
 from dcos_e2e.base_classes import ClusterBackend
 from dcos_e2e.cluster import Cluster
-from dcos_e2e.docker_versions import DockerVersion
 from dcos_e2e.node import Output, Role
 
 
@@ -21,9 +20,7 @@ class TestAdvancedInstallationMethod:
         """
         It is possible to install DC/OS on a node from a URL.
         """
-        # We use a specific version of Docker on the nodes because else we may
-        # hit https://github.com/opencontainers/runc/issues/1175.
-        cluster_backend = Docker(docker_version=DockerVersion.v17_12_1_ce)
+        cluster_backend = Docker()
         with Cluster(cluster_backend=cluster_backend) as cluster:
             for nodes, role in (
                 (cluster.masters, Role.MASTER),
@@ -44,9 +41,7 @@ class TestAdvancedInstallationMethod:
         """
         It is possible to install DC/OS on a node from a path.
         """
-        # We use a specific version of Docker on the nodes because else we may
-        # hit https://github.com/opencontainers/runc/issues/1175.
-        cluster_backend = Docker(docker_version=DockerVersion.v17_12_1_ce)
+        cluster_backend = Docker()
         with Cluster(cluster_backend=cluster_backend) as cluster:
             for nodes, role in (
                 (cluster.masters, Role.MASTER),